Refactor client module, add convenient client objects to do REST operations
