+++ src/main/js/index.js
@@ -3,7 +3,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
-import client from './components/client'
+import {client} from './client'
 import Home from './components/Home'
 import CreateUser from './components/CreateUser'
 import UpdateUser from './components/UpdateUser'
